[not verified] update modal style
