+++ projects/plugins/jetpack/extensions/plugins/launchpad-post-save-modal/editor.js
