+++ projects/plugins/jetpack/extensions/plugins/launchpad-post-save-modal/launchpad-post-save-modal.php
@@ -6,23 +6,23 @@
  **/
 
 
-	// Populate the available extensions with test-plugin.
+	// Populate the available extensions with launchpad-post-save-modal.
 	add_filter(
 		'jetpack_set_available_extensions',
 		function ( $extensions ) {
 			return array_merge(
 				$extensions,
 				array(
-					'test-plugin',
+					'launchpad-post-save-modal',
 				)
 			);
 		}
 	);
 
-	// Set the test-plugin availability, depending on the site plan.
+	// Set the launchpad-post-save-modal availability, depending on the site plan.
 	add_action(
 		'jetpack_register_gutenberg_extensions',
 		function () {
-			\Jetpack_Gutenberg::set_availability_for_plan( 'test-plugin' );
+			\Jetpack_Gutenberg::set_availability_for_plan( 'launchpad-post-save-modal' );
 		}
 	);
